@@ -30,29 +30,29 @@
   "devDependencies": {
     "autoprefixer": "7.2.5",
     "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.0",
-    "babel-eslint": "^8.2.1",
+    "babel-core": "^6.26.3",
+    "babel-eslint": "^8.2.6",
     "babel-loader": "^6.4.1",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.6.1",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
-    "babel-preset-react-app": "^3.1.1",
+    "babel-preset-react-app": "^3.1.2",
     "big.js": "^5.0.3",
     "chalk": "^2.3.0",
     "cross-spawn": "^5.1.0",
-    "css-loader": "^0.28.9",
+    "css-loader": "^0.28.11",
     "dotenv": "^4.0.0",
-    "eslint": "^4.15.0",
+    "eslint": "^4.19.1",
     "eslint-config-react-app": "^2.1.0",
     "eslint-loader": "^1.9.0",
     "eslint-plugin-flowtype": "^2.41.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.5.1",
-    "express": "^4.16.2",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.14.3",
+    "express": "^4.17.1",
     "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^1.1.6",
-    "filesize": "^3.5.11",
+    "file-loader": "^1.1.11",
+    "filesize": "^3.6.1",
     "fs-extra": "^5.0.0",
     "gzip-size": "^4.1.0",
     "gzippo": "^0.2.0",
@@ -61,16 +61,16 @@
     "json-loader": "^0.5.7",
     "morgan": "^1.9.0",
     "object-assign": "^4.1.1",
-    "opn": "^5.2.0",
+    "opn": "^5.5.0",
     "path-exists": "^3.0.0",
     "postcss-loader": "3.0.0",
-    "promise": "^8.0.1",
-    "prop-types": "^15.6.0",
-    "react": "^16.2.0",
-    "react-dev-utils": "^7.0.1",
-    "react-dom": "^16.2.0",
+    "promise": "^8.0.3",
+    "prop-types": "^15.7.2",
+    "react": "^16.9.0",
+    "react-dev-utils": "^7.0.5",
+    "react-dom": "^16.9.0",
     "react-syntax-highlighter": "^6.1.2",
-    "recursive-readdir": "^2.2.1",
+    "recursive-readdir": "^2.2.2",
     "rimraf": "^2.6.2",
     "strip-ansi": "^4.0.0",
     "style-loader": "^0.19.1",
